simplify context switching on aarch64
- use always an exception to trigger a context switch
